refactor resolvers
