+++ api/package.json
@@ -37,7 +37,10 @@
     "winston": "^3.1.0"
   },
   "devDependencies": {
+    "@babel/plugin-proposal-class-properties": "^7.1.0",
+    "@babel/plugin-proposal-throw-expressions": "^7.0.0",
     "graphql-cli": "^2.16.5",
+    "lodash.merge": "^4.6.1",
     "npm-run-all": "4.1.3",
     "prisma": "^1.18.1"
   }
